Fix CS8604 warnings
