+++ StyleChecker/StyleChecker/Refactoring/TypeClassParameter/Documents.cs
@@ -1,0 +1,44 @@
+﻿namespace StyleChecker.Refactoring.TypeClassParameter;
+
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis;
+
+/// <summary>
+/// Provides utility methods for working with documents.
+/// </summary>
+public static class Documents
+{
+    /// <summary>
+    /// Retrieves the semantic model, parameter symbol, and method symbol for
+    /// the specified document and syntax node.
+    /// </summary>
+    /// <param name="document">
+    /// The document to retrieve symbols from.
+    /// </param>
+    /// <param name="token">
+    /// The cancellation token.
+    /// </param>
+    /// <param name="node">
+    /// The syntax node to retrieve symbols for.
+    /// </param>
+    /// <returns>
+    /// A tuple containing the semantic model, parameter symbol, and method
+    /// symbol if found; otherwise, null.
+    /// </returns>
+    public static async Task<(
+            SemanticModel Model,
+            ISymbol Parameter,
+            IMethodSymbol Method)?>
+        GetSymbols(Document document, CancellationToken token, SyntaxNode node)
+    {
+        if (await document.GetSemanticModelAsync(token)
+                .ConfigureAwait(false) is not {} model
+            || model.GetDeclaredSymbol(node, token) is not {} parameter
+            || parameter.ContainingSymbol is not IMethodSymbol method)
+        {
+            return null;
+        }
+        return (model, parameter, method);
+    }
+}
